move parallelbigrnn into api
